INTERLOK-3828 Beef up the file resolver
* Reduce down to one file (meta)data expression format
* Allow paths to be further resovled
* Add unit test
