+++ interlok-common/src/main/java/com/adaptris/interlok/resolver/FileResolver.java
@@ -2,8 +2,10 @@ package com.adaptris.interlok.resolver;
 
 import com.adaptris.annotation.AdvancedConfig;
 import com.adaptris.annotation.InputFieldDefault;
+import com.adaptris.interlok.types.InterlokMessage;
 import lombok.Getter;
 import lombok.Setter;
+import org.apache.commons.lang3.StringUtils;
 
 import javax.validation.Valid;
 import java.io.File;
@@ -24,27 +26,28 @@ import java.util.regex.Pattern;
  * local file system.
  * <p>
  * This resolver resolves values based on the given path:
- * %file{what:path}, where 'what' can be one of several things to
+ * %file{path:%what...}, where '%what' can be one of several things to
  * resolve:
- *  - data        - The actual file data (currently only supports text
+ *  - %data        - The actual file data (currently only supports text
  *                  files as the resolve method returns a String).
- *  - size        - The file size (not meaningful for directories or
+ *  - %size        - The file size (not meaningful for directories or
  *                  other non-regular files).
- *  - type        - The type of file (regular file, directory, symlink,
+ *  - %type        - The type of file (regular file, directory, symlink,
  *                  etc).
- *  - date_create - The date of file creation.
- *  - date_modify - The last time the files was modified.
- *  - date_access - The date the file was last accessed.
- *  - permissions - The file permissions (Unix-like; ugo-rwx).
+ *  - %date_create - The date of file creation.
+ *  - %date_modify - The last time the files was modified.
+ *  - %date_access - The date the file was last accessed.
+ *  - %permissions - The file permissions (Unix-like; ugo-rwx).
  * </p>
  */
 public class FileResolver extends ResolverImp
 {
-	// Should match %file{data:...}$file{size:...} etc.
-	private static final String RESOLVE_REGEXP = "^.*%file\\{([\\w]+):([^:]+)\\}.*$";
+	// Should match %file{file:%data%size%...}
+	private static final String RESOLVE_REGEXP = "^.*%file\\{([^:]+):([%\\w]+)\\}.*$";
 	private transient Pattern resolverPattern;
 
 	private static final String DATE_FORMAT = "yyyy-MM-dd HH:mm:ss zzz";
+	private static final String SEPARATOR = ",";
 
 	@Getter
 	@Setter
@@ -52,6 +55,13 @@ public class FileResolver extends ResolverImp
 	@InputFieldDefault(DATE_FORMAT)
 	@AdvancedConfig(rare = true)
 	private String dateFormat;
+
+	@Getter
+	@Setter
+	@Valid
+	@InputFieldDefault(SEPARATOR)
+	@AdvancedConfig(rare = true)
+	private String separator;
 
 	public FileResolver()
 	{
@@ -67,6 +77,12 @@ public class FileResolver extends ResolverImp
 	@Override
 	public String resolve(String lookupValue)
 	{
+		return resolve(lookupValue, (InterlokMessage)null);
+	}
+
+	@Override
+	public String resolve(String lookupValue, InterlokMessage target)
+	{
 		if (lookupValue == null)
 		{
 			return null;
@@ -76,32 +92,57 @@ public class FileResolver extends ResolverImp
 		Matcher m = resolverPattern.matcher(lookupValue);
 		while (m.matches())
 		{
-			What what = What.parse(m.group(1));
-			File path = new File(m.group(2));
-			log.trace("Resolve {} on path {} ", what, path);
-			if (what == null)
+			File path;
+			String pathReplace = m.group(1);
+			// see if path is resolvable...
+			if (target != null)
 			{
-				log.error("{} is not something that's resolvable", m.group(1));
-				throw new UnresolvableException(m.group(1) + " is not something that's resolvable");
+				path = new File(target.resolve(pathReplace));
+			}
+			else
+			{
+				path = new File(pathReplace);
 			}
 
-			String value = "";
-			try
+			String w = m.group(2);
+			StringBuffer sb = new StringBuffer();
+			for (String s : w.split("%"))
 			{
-				value = what.resolve(path.toPath());
-				if (value == null)
+				if (StringUtils.isEmpty(s))
 				{
-					log.warn("{} resolved to null for {}", what, path);
-					value = "";
+					continue;
 				}
-			}
-			catch (IOException e)
-			{
-				log.error("Could not resolve {} on path {} ", what, path, e);
+				What what = What.parse(s);
+				if (what == null)
+				{
+					log.error("{} is not something that's resolvable", s);
+					throw new UnresolvableException(s + " is not something that's resolvable");
+				}
+				log.trace("Resolve {} on path {} ", what, path);
+
+				String value = "";
+				try
+				{
+					value = what.resolve(path.toPath());
+					if (value == null)
+					{
+						log.warn("{} resolved to null for {}", what, path);
+						value = "";
+					}
+				}
+				catch (IOException e)
+				{
+					log.error("Could not resolve {} on path {} ", what, path, e);
+				}
+				if (sb.length() > 0 && value.length() > 0)
+				{
+					sb.append(separator());
+				}
+				sb.append(value);
 			}
 
-			String toReplace = "%file{" + what + ":" + path + "}";
-			result = result.replace(toReplace, value);
+			String toReplace = "%file{" + pathReplace + ":" + w + "}";
+			result = result.replace(toReplace, sb.toString());
 			m = resolverPattern.matcher(result);
 		}
 		return result;
@@ -117,6 +158,11 @@ public class FileResolver extends ResolverImp
 	public boolean canHandle(String value)
 	{
 		return resolverPattern.matcher(value).matches();
+	}
+
+	private String separator()
+	{
+		return StringUtils.defaultIfEmpty(separator, SEPARATOR);
 	}
 
 	public enum What
