+++ interlok-common/src/main/resources/META-INF/services/com.adaptris.interlok.resolver.Resolver
@@ -1,4 +1,4 @@
 com.adaptris.interlok.resolver.FromEnvironment
 com.adaptris.interlok.resolver.FromSystemProperties
 com.adaptris.interlok.resolver.FileResolver
-com.adaptris.interlok.resolver.AltFileResolver
+com.adaptris.interlok.resolver.FileResolver
